Commit: TC_09_06 fix test and page structure.
