+++ pom/pages/HelpContactUsPage.js
@@ -8,90 +8,6 @@ export const CONTACT_US_INPUT_NAME = {
     subject: 'emailSubject',
 };
 
-export const CONTACT_US_DROPDOWN = [
-    {
-        name: 'Account Management & Access',
-        subcategories: ['Password Reset', 'Credentials/Username Issues', '2FA Issues', 'Other'],
-    },
-    {
-        name: 'Payment, Sales & Billing Issues',
-        subcategories: ['Order Status', 'Declined Payment', 'Refund', 'Invoice', 'Other'],
-    },
-    {
-        name: 'Domains',
-        subcategories: [
-            'Nameserver Changes',
-            'DNSSEC',
-            'DNS Records',
-            'Transfer to Trustname',
-            'Transfer from Trustname',
-            'Transfer Disputes',
-            'Domain Contact Issue',
-            'Whois Verification',
-            'Web Forwarding',
-            'Renewal',
-            'Other',
-        ],
-    },
-    {
-        name: 'Pre-Sales',
-        subcategories: ['Domains', 'Hosting', 'SSL', 'Other'],
-    },
-    {
-        name: 'Hosting',
-        subcategories: [
-            'Database',
-            'File Transfer',
-            'Server Creation',
-            'VPS & Dedicated Servers',
-            'Blocked Server',
-            'Attach to a Domain',
-            'Change the Owner',
-            'Renewal, Credentials/Access Issues',
-            'Other',
-        ],
-    },
-    {
-        name: 'SSL',
-        subcategories: ['Revocation', 'Other'],
-    },
-    {
-        name: 'Private Email',
-        subcategories: [
-            'Mailbox Suspended',
-            'Configuration',
-            'Problems Sending Emails',
-            'Problems Receiving Emails',
-            'Suspicious Phishing/Hacked Mails',
-            'Other',
-        ],
-    },
-    {
-        name: 'Affiliate',
-        subcategories: [],
-    },
-    {
-        name: 'Bug/Technical Issue',
-        subcategories: [],
-    },
-    {
-        name: 'Feature/Product Request',
-        subcategories: [],
-    },
-    {
-        name: 'Commercial Matters',
-        subcategories: [],
-    },
-    {
-        name: 'Media Inquiries',
-        subcategories: [],
-    },
-    {
-        name: 'Feedback',
-        subcategories: [],
-    },
-];
-
 export default class HelpContactUsPage {
     constructor(page) {
         this.page = page;
@@ -242,13 +158,9 @@ export default class HelpContactUsPage {
         });
     }
 
-    async verifyDropdownItems(type) {
+    async verifyDropdownItems(type, subItems) {
         await step(`Verify is the list of dropdown items corresponds to the "${type}".`, async () => {
             let dropdownItems = await this.items.allInnerTexts();
-            let subItems =
-                type === 'Type'
-                    ? CONTACT_US_DROPDOWN.map((c) => c.name)
-                    : CONTACT_US_DROPDOWN.find((c) => c.name === type).subcategories;
             let dropdownItemsSorted = dropdownItems.sort();
             let subItemsSorted = subItems.sort();
             expect(
